backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor==6.3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-storages==1.13.1
djangorestframework==3.14.0
pillow==11.1.0
psycopg2==2.9.10
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
